Songs-backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
django-request==1.6.2
djangorestframework==3.14.0
importlib-metadata==6.2.1
Markdown==3.4.1
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
zipp==3.15.0
